extracts base cache sql client to create views on any destination
