+++ tests/load/filesystem/test_filesystem_sql_secrets.py
@@ -60,8 +60,8 @@ def test_secrets_management(
     ) -> FilesystemSqlClient:
         return FilesystemSqlClient(
             dataset_name="second",
-            fs_client=pipeline.destination_client(),  #  type: ignore
-            credentials=DuckDbCredentials(connection),
+            remote_client=pipeline.destination_client(),  #  type: ignore
+            cache_db=DuckDbCredentials(connection),
         )
 
     def _secrets_exist() -> bool:
@@ -79,7 +79,11 @@ def test_secrets_management(
     external_db = _external_duckdb_connection()
     fs_sql_client = _fs_sql_client_for_external_db(external_db)
     with fs_sql_client as sql_client:
-        fs_sql_client.create_authentication(persistent=True)
+        fs_sql_client.create_authentication(
+            fs_sql_client.remote_client.config.bucket_url,
+            fs_sql_client.remote_client.config.credentials,
+            persistent=True,
+        )
     assert _secrets_exist()
 
     # remove secrets and check that they are removed
@@ -91,11 +95,15 @@ def test_secrets_management(
     # prevent creating persistent secrets on in mem databases
     fs_sql_client = FilesystemSqlClient(
         dataset_name="second",
-        fs_client=pipeline.destination_client(),  #  type: ignore
+        remote_client=pipeline.destination_client(),  #  type: ignore
     )
     with pytest.raises(Exception):
         with fs_sql_client as sql_client:
-            fs_sql_client.create_authentication(persistent=True)
+            fs_sql_client.create_authentication(
+                fs_sql_client.remote_client.config.bucket_url,
+                fs_sql_client.remote_client.config.credentials,
+                persistent=True,
+            )
 
     # check that no warning was logged
     logger_spy.assert_not_called()
@@ -106,7 +114,11 @@ def test_secrets_management(
     duck_db = duckdb.connect(duck_db_location)
     fs_sql_client = _fs_sql_client_for_external_db(duck_db)
     with fs_sql_client as sql_client:
-        sql_client.create_authentication(persistent=True)
+        sql_client.create_authentication(
+            fs_sql_client.remote_client.config.bucket_url,
+            fs_sql_client.remote_client.config.credentials,
+            persistent=True,
+        )
     logger_spy.assert_called_once()
     assert warning_mesage in logger_spy.call_args_list[0][0][0]
     duck_db.close()
